+++ pkg/binder/binding/resourcereservation/resource_reservation.go
@@ -47,17 +47,19 @@ const (
 )
 
 type service struct {
-	fakeGPuNodes        bool
-	kubeClient          client.WithWatch
-	reservationPodImage string
-	allocationTimeout   time.Duration
-	gpuGroupMutex       *group_mutex.GroupMutex
-	namespace           string
-	serviceAccountName  string
-	appLabelValue       string
-	scalingPodNamespace string
-	runtimeClassName    string
-	podResources        *v1.ResourceRequirements
+	fakeGPuNodes                        bool
+	kubeClient                          client.WithWatch
+	reservationPodImage                 string
+	allocationTimeout                   time.Duration
+	gpuGroupMutex                       *group_mutex.GroupMutex
+	namespace                           string
+	serviceAccountName                  string
+	appLabelValue                       string
+	scalingPodNamespace                 string
+	runtimeClassName                    string
+	podResources                        *v1.ResourceRequirements
+	reservationPodSecurityContext       *v1.PodSecurityContext
+	reservationContainerSecurityContext *v1.SecurityContext
 }
 
 func NewService(
@@ -71,19 +73,23 @@ func NewService(
 	scalingPodNamespace string,
 	runtimeClassName string,
 	podResources *v1.ResourceRequirements,
+	reservationPodSecurityContext *v1.PodSecurityContext,
+	reservationContainerSecurityContext *v1.SecurityContext,
 ) *service {
 	return &service{
-		fakeGPuNodes:        fakeGPuNodes,
-		kubeClient:          kubeClient,
-		reservationPodImage: reservationPodImage,
-		allocationTimeout:   allocationTimeout,
-		gpuGroupMutex:       group_mutex.NewGroupMutex(),
-		namespace:           namespace,
-		serviceAccountName:  serviceAccountName,
-		appLabelValue:       appLabelValue,
-		scalingPodNamespace: scalingPodNamespace,
-		runtimeClassName:    runtimeClassName,
-		podResources:        podResources,
+		fakeGPuNodes:                        fakeGPuNodes,
+		kubeClient:                          kubeClient,
+		reservationPodImage:                 reservationPodImage,
+		allocationTimeout:                   allocationTimeout,
+		gpuGroupMutex:                       group_mutex.NewGroupMutex(),
+		namespace:                           namespace,
+		serviceAccountName:                  serviceAccountName,
+		appLabelValue:                       appLabelValue,
+		scalingPodNamespace:                 scalingPodNamespace,
+		runtimeClassName:                    runtimeClassName,
+		podResources:                        podResources,
+		reservationPodSecurityContext:       reservationPodSecurityContext,
+		reservationContainerSecurityContext: reservationContainerSecurityContext,
 	}
 }
 
@@ -532,6 +538,7 @@ func (rsc *service) createResourceReservationPod(
 				}
 				return &rsc.runtimeClassName
 			}(),
+			SecurityContext:    rsc.reservationPodSecurityContext,
 			ServiceAccountName: rsc.serviceAccountName,
 			Containers: []v1.Container{
 				{
@@ -539,6 +546,7 @@ func (rsc *service) createResourceReservationPod(
 					Image:           rsc.reservationPodImage,
 					ImagePullPolicy: v1.PullIfNotPresent,
 					Resources:       resources,
+					SecurityContext: rsc.reservationContainerSecurityContext,
 					Env: []v1.EnvVar{
 						{
 							Name: "POD_NAME",